# Improved sorting structure and prepared for release 0.1.2